feat: application layer structure

+++ TraderForge.Domain/Entities/MarketAsset.cs
@@ -5,6 +5,6 @@ public class MarketAsset
     public int Id { get; set; }
     public string Symbol { get; set; }
     public string Name { get; set; }
-    public float CurrentPrice { get; set; }
+    public decimal CurrentPrice { get; set; }
     public DateTime LastUpdated { get; set; }
 }
